+++ src/Controller/AbstractTemplateController.php
@@ -2,19 +2,19 @@
 
 declare(strict_types=1);
 
-namespace App\Repositories;
+namespace App\Controller;
 
 use Twig\Environment;
 use Twig\Loader\FilesystemLoader;
 
-abstract class TemplateRepository
+abstract class AbstractTemplateController
 {
     private $loader;
     protected $twig;
     protected $template = 'main';
     public function __construct(string $template)
     {
-        $this->loader = new FilesystemLoader('src/View/Templates');
+        $this->loader = new FilesystemLoader('src/View/Template');
         $this->twig = new Environment($this->loader);
         $this->template = $template;
     }