+++ src/Lib/common.php
@@ -191,7 +191,7 @@ function post($k)
     return Ses::$r['post'][$k] ?? '';
 }
 
-#services
+#Service
 
 //ses
 function voc(string $d): string